+++ CHANGELOG.md
@@ -2,11 +2,20 @@
 
 All notable changes to the "todo-tools" extension will be documented in this file.
 
+## [0.7.3] - 2022-06-27
+### add
+- priority setting commands
+- due date +7 command
+- keybindings
+
+## [0.7.2] - 2022-06-01
+### add
+- sort @started items to the top
+
 ## [0.7.1] - 2022-05-26
 ### fix
 - future recurring items not moving to future
-- 'today' day not being reevaluated as time PASSES
-
+- 'today' day not being reevaluated as time passes
 
 ## [0.7.0] - 2022-05-25
 ### add